Refactor workspace component folder location
